@@ -1,28 +1,28 @@
-<?php declare(strict_types=1);
+<?php
 
-/*
- * This file is part of the PsyshBundle package.
- *
- * (c) Théo FIDRY <theo.fidry@gmail.com>
+/**
+ * @copyright Théo FIDRY <theo.fidry@gmail.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Fidry\PsyshBundle\Command;
+declare(strict_types=1);
 
+namespace TeamDotBlue\PsyshBundle\Test\Command;
+
+use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\TestCase;
 use Psy\Shell;
 use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
+
+use TeamDotBlue\PsyshBundle\Command\PsyshCommand;
+
 use function is_a;
 
-/**
- * @covers \Fidry\PsyshBundle\Command\PsyshCommand
- *
- * @author Théo FIDRY <theo.fidry@gmail.com>
- */
+#[CoversClass(PsyshCommand::class)]
 class PsyshCommandUnitTest extends TestCase
 {
     public function testIsASymfonyCommand(): void
