@@ -1,40 +1,30 @@
-<?php declare(strict_types=1);
+<?php
 
-/*
- * This file is part of the PsyshBundle package.
- *
- * (c) Théo FIDRY <theo.fidry@gmail.com>
+/**
+ * @copyright Théo FIDRY <theo.fidry@gmail.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Fidry\PsyshBundle;
+declare(strict_types=1);
 
+namespace TeamDotBlue\PsyshBundle\Test;
+
+use PHPUnit\Framework\Attributes\CoversNothing;
 use ReflectionClass;
 use Symfony\Bundle\FrameworkBundle\Test\KernelTestCase;
+use TeamDotBlue\PsyshBundle\PsyshFacade;
 
-/**
- * @coversNothing
- *
- * @author Théo FIDRY <theo.fidry@gmail.com>
- */
+#[CoversNothing]
 class PsyshFacadeTest extends KernelTestCase
 {
-    /**
-     * Test that the bundle loads and compiles.
-     */
     public function testServicesLoading(): void
     {
         static::bootKernel();
-
         PsyshFacade::init();
-
-        $shellReflection= (new ReflectionClass(PsyshFacade::class))->getProperty('shell');
+        $shellReflection = (new ReflectionClass(PsyshFacade::class))->getProperty('shell');
         $shellReflection->setAccessible(true);
-
-        $shell = $shellReflection->getValue();
-
-        $this->assertNotNull($shell);
+        $this->assertNotNull($shellReflection->getValue());
     }
 }
