@@ -1,26 +1,23 @@
-<?php declare(strict_types=1);
+<?php
 
-/*
- * This file is part of the PsyshBundle package.
- *
- * (c) Théo FIDRY <theo.fidry@gmail.com>
+/**
+ * @copyright Théo FIDRY <theo.fidry@gmail.com>
  *
  * For the full copyright and license information, please view the LICENSE
  * file that was distributed with this source code.
  */
 
-namespace Fidry\PsyshBundle;
+declare(strict_types=1);
+
+namespace TeamDotBlue\PsyshBundle;
 
 use Psy\Shell;
 use RuntimeException;
-use Symfony\Component\DependencyInjection\ContainerAwareInterface;
 use Symfony\Component\DependencyInjection\ContainerInterface;
+
 use function array_merge;
 
-/**
- * @author Théo FIDRY <theo.fidry@gmail.com>
- */
-final class PsyshFacade implements ContainerAwareInterface
+final class PsyshFacade
 {
     private static Shell $shell;
 
@@ -36,16 +33,19 @@ final class PsyshFacade implements ContainerAwareInterface
             throw new RuntimeException('Cannot initialize the facade without a container.');
         }
 
-        self::$shell = self::$container->get('psysh.shell');
+        self::$shell = self::$container->get(Shell::class);
     }
 
+    /**
+     * @deprecated Due to {@see Shell::debug} being deprecated
+     *
+     * @param array<mixed> $variables
+     * @param object|string|callable $bind
+     */
     public static function debug(array $variables = [], $bind = null): void
     {
         self::init();
-
-        $_variables = array_merge(self::$shell->getScopeVariables(), $variables);
-
-        self::$shell::debug($_variables, $bind);
+        self::$shell::debug(array_merge(self::$shell->getScopeVariables(), $variables), $bind);
     }
 
     public function setContainer(ContainerInterface $container = null): void
